Split pricing routes out of channelPointsAdminMutations to fix qlty file-complexity flag
Moves the pricing-config, pricing-delete, and pricing-settings routes (plus
the cooldown-resolution helper) into a new channelPointsAdminPricingMutations.ts,
leaving channelPointsAdminMutations.ts scoped to reward CRUD. The shared
reward-delete-action helper moves to channelPointsAdminShared.ts so both
route files can use it. No behavior change.
