+++ src/web/routes/channelPointsAdminShared.test.ts
@@ -6,6 +6,7 @@ import { getStreamerByDiscordId } from '../../db';
 import {
   requireStreamer, parsePositiveIntField, parseNonNegativeNumberField, parsePositiveNumberField,
   parseCheckboxField, parseHexColorField, parseRewardIdParam, parseRewardFields, effectiveCooldownSeconds,
+  handleRewardDeleteAction,
 } from './channelPointsAdminShared';
 
 const VALID_REWARD_ID = '12345678-1234-1234-8234-123456789abc';
@@ -37,6 +38,52 @@ describe('requireStreamer', () => {
     const result = await requireStreamer(req, res);
     expect(result).toBeNull();
     expect(res.redirect).toHaveBeenCalledWith('/channel-points?error=not_a_streamer');
+  });
+});
+
+describe('handleRewardDeleteAction', () => {
+  const streamer = { id: 1 };
+  function makeReqRes(twitchRewardId = VALID_REWARD_ID, discordId = '123') {
+    const req = { session: { user: { discordId } }, params: { twitchRewardId } } as any;
+    const res = { redirect: vi.fn() } as any;
+    return { req, res };
+  }
+  const opts = { log: { error: vi.fn() } as any, successCode: 'thing_deleted', errorLogLabel: 'delete error:', errorCode: 'delete_failed' };
+
+  it('redirects to not_a_streamer without running the action when the requester is not a streamer', async () => {
+    vi.mocked(getStreamerByDiscordId).mockResolvedValue(null);
+    const action = vi.fn();
+    const { req, res } = makeReqRes();
+    await handleRewardDeleteAction(req, res, action, opts);
+    expect(action).not.toHaveBeenCalled();
+    expect(res.redirect).toHaveBeenCalledWith('/channel-points?error=not_a_streamer');
+  });
+
+  it('redirects to invalid_reward_id without running the action for a malformed reward id', async () => {
+    vi.mocked(getStreamerByDiscordId).mockResolvedValue(streamer as any);
+    const action = vi.fn();
+    const { req, res } = makeReqRes('not-a-uuid');
+    await handleRewardDeleteAction(req, res, action, opts);
+    expect(action).not.toHaveBeenCalled();
+    expect(res.redirect).toHaveBeenCalledWith('/channel-points?error=invalid_reward_id');
+  });
+
+  it('runs the action and redirects to the success code', async () => {
+    vi.mocked(getStreamerByDiscordId).mockResolvedValue(streamer as any);
+    const action = vi.fn().mockResolvedValue(undefined);
+    const { req, res } = makeReqRes();
+    await handleRewardDeleteAction(req, res, action, opts);
+    expect(action).toHaveBeenCalledWith(streamer, VALID_REWARD_ID);
+    expect(res.redirect).toHaveBeenCalledWith('/channel-points?success=thing_deleted');
+  });
+
+  it('logs and redirects to the error code when the action throws', async () => {
+    vi.mocked(getStreamerByDiscordId).mockResolvedValue(streamer as any);
+    const action = vi.fn().mockRejectedValue(new Error('boom'));
+    const { req, res } = makeReqRes();
+    await handleRewardDeleteAction(req, res, action, opts);
+    expect(opts.log.error).toHaveBeenCalledWith(opts.errorLogLabel, expect.any(Error));
+    expect(res.redirect).toHaveBeenCalledWith('/channel-points?error=delete_failed');
   });
 });
